Banco CDIS Completo

+++ bancoCdisAABC/src/main/java/com/bancoCdisAABC/springoot/app/controller/TarjetaController.java
@@ -1,5 +1,6 @@
 package com.bancoCdisAABC.springoot.app.controller;
 
+import java.util.List;
 import java.util.Map;
 
 import javax.validation.Valid;
@@ -8,27 +9,39 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
+import org.springframework.web.bind.WebDataBinder;
+import org.springframework.web.bind.annotation.InitBinder;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.SessionAttributes;
 import org.springframework.web.bind.support.SessionStatus;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import com.bancoCdisAABC.springoot.app.editors.CuentaPropetyEditor;
+import com.bancoCdisAABC.springoot.app.errors.DataBaseBancoException;
 import com.bancoCdisAABC.springoot.app.models.dao.ICuentaDao;
 import com.bancoCdisAABC.springoot.app.models.dao.ITarjetaDao;
+import com.bancoCdisAABC.springoot.app.models.entity.Cuenta;
 import com.bancoCdisAABC.springoot.app.models.entity.Tarjeta;
 
 @Controller
+@SessionAttributes("tarjeta")
 public class TarjetaController {
 
 	@Autowired
 	private ITarjetaDao tarjetaDao;
-	
+
 	@Autowired
 	private ICuentaDao cuentaDao;
 
 	@Autowired
 	private CuentaPropetyEditor cuentaEditor;
+	
+	@InitBinder
+	public void InitBinder(WebDataBinder binder) {
+		binder.registerCustomEditor(Cuenta.class, "cuenta", cuentaEditor);
+	}
 	
 	@RequestMapping(value = "/tarjetas-lista", method = RequestMethod.GET)
 	public String listar(Model model) {
@@ -38,9 +51,11 @@ public class TarjetaController {
 	}
 		
 		@RequestMapping(value = "/formatarjeta")
-		public String crear(Map<String, Object> model ) {
+		public String crear(Map<String, Object> model, Model modelList ) {
 			Tarjeta tarjeta = new Tarjeta();
+			List<Cuenta> listaCuentas = cuentaDao.findAll();
 			model.put("tajeta", tarjeta );
+			modelList.addAttribute("listaCuentas", listaCuentas);
 			model.put("titulo", "Llenar los datos de una tarjeta" );
 			return "formtarjeta";
 	}
@@ -49,7 +64,7 @@ public class TarjetaController {
 		public String editar(@PathVariable(value = "id") Long id, Map<String, Object>model) {
 			Tarjeta tarjeta = null;
 			
-			if(id > 0) {
+			if(id > 0 && id!= null) {
 				tarjeta = tarjetaDao.findOne(id);
 			}else {
 				return "redirect:/index";
@@ -61,28 +76,38 @@ public class TarjetaController {
 		}
 		
 		@RequestMapping(value = "/formtarjeta", method = RequestMethod.POST)
-		public String guardar(@Valid Tarjeta tarjeta, BindingResult result, Model model, SessionStatus status) {
+		public String guardar(@Valid Tarjeta tarjeta, BindingResult result, Model model, SessionStatus status,
+				RedirectAttributes flash) {
 			
 			if(result.hasErrors()) {
-				model.addAttribute("titulo", "Forlumario de tarjeta");
+				model.addAttribute("titulo", "Llene correctamente los cargos");
+				model.addAttribute("result", result.hasErrors());
+				model.addAttribute("mensaje", "Error al enviar los datos, por favor escriba correctamente los campos");
 				return "formtarjeta";
+			}else {
+				model.addAttribute("result", false);
 			}
-			tarjetaDao.save(tarjeta);
+			
+			model.addAttribute("titulo", "Formulario de tarjeta");
+			model.addAttribute("mensaje", "Se envio la informacion correctamente");
+			try {
+				tarjetaDao.save(tarjeta);
+			} catch (DataBaseBancoException e) {
+				e.printStackTrace();
+				flash.addFlashAttribute("mensaje", e.getMessage());
+			}
 			status.setComplete();
-			
-			return "redirect:index";
+
+			return "redirect:formtarjeta";
 		}
-		
-		@RequestMapping(value="/eliminar/{id}")
+		@RequestMapping(value = "/eliminar/{id}")
 		public String eliminar(@PathVariable(value = "id") Long id) {
-			
-			if(id > 0) {
+
+			if (id != null && id > 0) {
 				tarjetaDao.delete(id);
 			}
-			return "redirect:index";		
+			return "redirect:/tarjetas-lista";
 		}
-		
-		@RequestMapping(value="/eliminar/{id}")
 	
 	
 }
